requirements.txt
azure-ai-documentintelligence
openai
pandas
openpyxl
streamlit
python-dotenv